feat: 홈 객체 생성 및 메서도 이전
