+++ src/main/java/com/ryan/roomreservationservice/utils/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.ryan.roomreservationservice.domain.utils.exception;
+package com.ryan.roomreservationservice.utils.exception;
 
 public class ErrorMessage {
     public static String UNAVAILABLE_RESERVATION = "예약이 불가능합니다.";
